Removed matching bracket open/close keybinding
smart-parens replaced this and "M-[" caused bracketing paste mode to
fail. The reason
is (https://emacs.stackexchange.com/questions/28851/how-to-turn-off-bracketed-paste-mode):

"My crystal ball said you should remove
(global-set-key (kbd "M-[") 'previous-multiframe-window)
from your ~/.emacs. You might like to M-x report-emacs-bug and ask for
global-set-key to complain when the key sequence conflicts with
something in input-decode-map (as is the case here, because the byte
sequence ESC [ (which corresponds to the M-[ event) is used by text
terminal as a prefix for encoding all kinds of special events)."